requirements.txt
django
django-rest-framework
django-cors-headers
django-allauth
djangorestframework-simplejwt
pytest-django
Pillow